requirements.txt
langchain
groq
langchain_community
python-dotenv
langchain_core
